requirements.txt
duckduckgo_search==7.1.1
groq==0.13.1
phidata==2.7.5
yfinance==0.2.51
uv==0.5.13
